[BPK-4276] Restructure H and I component pages
