+++ docs/src/pages/HorizontalGridPage/HorizontalGridPage.js
@@ -17,68 +17,11 @@
  */
 
 import React from 'react';
-import BpkBlockquote from 'bpk-component-blockquote';
-import BpkRouterLink from 'bpk-component-router-link';
-import BpkText, { WEIGHT_STYLES } from 'bpk-component-text';
-import gridReadme from 'bpk-component-grid/README.md';
 
-import * as routes from '../../constants/routes';
-import { WebComponentPage } from '../../components/ComponentPage';
 import DocsPageWrapper from '../../components/DocsPageWrapper';
-import BpkParagraph from '../../components/Paragraph';
 import IntroBlurb from '../../components/IntroBlurb';
-import Code from '../../components/Code';
 
-const components = [
-  {
-    id: 'default',
-    title: 'Default',
-    examples: [
-      <BpkText tagName="h3" textStyle="base" weight={WEIGHT_STYLES.bold}>
-        Columns
-      </BpkText>,
-      <BpkParagraph>
-        The grid makes use of 12 percentage based columns, which means your
-        layouts will scale no matter the browser size. Columns are separated by
-        fixed width gutters and margins.
-      </BpkParagraph>,
-      <BpkParagraph>
-        <BpkRouterLink to={routes.GRID_COLUMN_DEMO}>View demo.</BpkRouterLink>
-      </BpkParagraph>,
-      <BpkText tagName="h3" textStyle="base" weight={WEIGHT_STYLES.bold}>
-        Offsets
-      </BpkText>,
-      <BpkParagraph>
-        Offsets allow you to shift columns to the right. Depending on their
-        size, columns can be offset by 1 through 11.
-      </BpkParagraph>,
-      <BpkParagraph>
-        <BpkRouterLink to={routes.GRID_OFFSET_DEMO}>View demo.</BpkRouterLink>
-      </BpkParagraph>,
-      <BpkText tagName="h3" textStyle="base" weight={WEIGHT_STYLES.bold}>
-        Responsive behaviour
-      </BpkText>,
-      <BpkParagraph>
-        The grid works in conjunction with the{' '}
-        {
-          <BpkRouterLink to={routes.BREAKPOINT}>
-            Breakpoint component
-          </BpkRouterLink>
-        }{' '}
-        and can be used to position content differently based on these viewports
-        by specifying different widths and offsets.
-      </BpkParagraph>,
-      <BpkBlockquote>
-        <strong>Note:</strong> The Backpack grid is intended to be used for
-        overall page layout as opposed to spacing out atom or molecule level
-        components. Please stick to flexbox based techniques for more intricate
-        layouts (or <Code>display: table;</Code> for browsers which lack
-        support), just be sure to use the spacing values above in order to
-        achieve consistency.
-      </BpkBlockquote>,
-    ],
-  },
-];
+import WebHorizontalGridPage from './WebHorizontalGridPage';
 
 const blurb = [
   <IntroBlurb>
@@ -86,20 +29,11 @@ const blurb = [
   </IntroBlurb>,
 ];
 
-const HorizontalGridSubPage = () => (
-  <WebComponentPage
-    examples={components}
-    readme={gridReadme}
-    sassdocId="grids"
-    packageName="bpk-component-grid"
-  />
-);
-
 const HorizontalGridPage = () => (
   <DocsPageWrapper
     title="Horizontal grid"
     blurb={blurb}
-    webSubpage={<HorizontalGridSubPage />}
+    webSubpage={<WebHorizontalGridPage />}
   />
 );
