+++ docs/src/pages/IconPage/AndroidIconPage.js
