+++ docs/src/pages/IconPage/WebIconsPage.js
