+++ docs/src/pages/ImagesPage/WebImagesPage.js
